+++ packages/plugins/LayerChooser/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## unpublished
+
+- Fix: The LayerChooser plugin will keep working on the error that a layer without entry in the service register has been configured.
+
 ## 1.2.0
 
 - Feature: Improved implementation to make plugin SPA-ready.