+++ packages/plugins/Legend/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## unpublished
+
+- Fix: The Legend plugin will keep working on the error that a layer without entry in the service register has been configured.
+
 ## 1.1.0
 
 - Feature: Locale string layer names are now translated.